% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\name{combinations}
\alias{combinations}
\title{combination function to obtain at random c3 matrices}
\usage{
combinations(number_of_time_segments)
}
\arguments{
\item{number_of_time_segments}{Numeric vector.}
}
\value{
list50 A list which contains a c3 matrix.
}
\description{
\code{combinations()} function to obtain at random c3 matrices for the construction of the vector timeMax  depending on the number of time segments
}
