% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab_E.R
\name{crosstab_E}
\alias{crosstab_E}
\title{crosstab function.
\code{crosstab function} Function to compute the group numbers of recreation area with the same activities on the same time segments or to compute the column created for the group numbers of recreation areas and the activities pattern on the same time segment.}
\usage{
crosstab_E(Rmatrix, Cmatrix, C1)
}
\arguments{
\item{Rmatrix}{Matrix of 1 or 2 columns:
The first column contains the vector RcrID or the vector groupsRecr_at and the second column if exist the vextor of the corresponding activities pattern for the vector groupsRecr_at}

\item{Cmatrix}{Matrix of 2 or 1 coloumn(s)
Activities patern ID and time segment for 2 columns or time segment for one column.}

\item{C1}{matrix empty,vector

If the third argument of the function C1 is an empty matrix and when the Rmatrix  matrix contains 1 column THEN the results of this function are the group numbers of recreation areas with the same activities and on the same time segments. If the Rmatrix  matrix contains 2 columns then the results of this function are the group numbers of groupsRecr_at and the activities pattern on the same time segment (see Wageningen_tables_Excel folder: Procedures_ACRER.xls, Excel sheet: crosstab_procedure (1 or 9) or HTML folder: Procedures_ACRER.html (crosstab_procedure (1 or 9) section)
If the third argument of the function C1 matrix is NOT an empty matrix but contains the linear optimization solution of our model THEN the result of this function are the linear optimization solutions per time segment.}
}
\value{
list1 Listrow_matrix, col_matrix, crosstable, groups_matrix, fdata, tdata.
}
\description{
crosstab_E(Rmatrix, Cmatrix, C1)
}
\details{
The data we use here are from the table data which we got from a query from a query of the Final table. The table data had n records (n rows) and 4 columns namely: recreation area, activities pattern, transport type and time segment.
This table data with the use of the function group_by was reduced in n1 records. The column of the trasport type is not important for the groups computation
The crosstab_E function has 5 subfunctions:
\enumerate{
\item NGroup0
\item NGroup1
\item NGroup2
\item NGroup3
\item NGroup4
}

NGroup0 computes the group numbers of recreation areas where all the sequence(s) of their presences (1) on the combination of activities pattern and time segment are not interrupted by absence (0). Thus in this case we have one group of ones (presence).
NGroup1 computes the group numbers of recreation areas where all the sequence(s) of their presences (1) on the combination of activities pattern and time segment are once interrupted by absence (0). Thus in this case we have two groups of ones (presence).
NGroup2 computes the group numbers of recreation areas where all the sequence(s) of their presences (1) on the combination of activities pattern and time segment are twice interrupted by absences (0). Thus in this case we have three groups of ones (presence).
NGroup3 computes the group numbers of recreation areas where all the sequence(s) of their presences (1) on the combination of activities pattern and time segment are three times interrupted by absences (0). Thus in this case we have four groups of ones (presence).
NGroup4 computes the group numbers of recreation areas where all the sequence(s) of their presences (1) on the combination of activities pattern and time segment are four times interrupted by absences (0). Thus in this case we have five groups of ones (presence)
}
