% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstallPackages.R
\name{InstallPackages}
\alias{InstallPackages}
\title{Install_Required_Packages
Create a function to check for installed packages and install them if they are not already installed.}
\usage{
InstallPackages(packages)
}
\arguments{
\item{packages}{The required packages}
}
\value{
InstallPackages Install the required packages
}
\description{
Install_Required_Packages
Create a function to check for installed packages and install them if they are not already installed.
}
