% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGroup4.R
\name{NGroup4}
\alias{NGroup4}
\title{NGroup4 computes the group numbers of recreation areas.}
\usage{
NGroup4(nRowMatrix, activ_segment, c1, c2, sumc1c2, sumc1, levels_sr,
  Index_group_number, groups, Nfactors, fRecrIDX, FRecrID, Ngroups,
  fbfactX1, fbfactX2, Fbfact11, Fbfact22, pos1, q)
}
\arguments{
\item{nRowMatrix}{Matrix.
Select each row of the cross table if the vector for the rows matrix equals to one.}

\item{activ_segment}{Matrix.
activ_segment construct of the presence/absence of activities-time segment matrix.}

\item{c1}{Numeric array.
Construct the row of the activities pattern factor in which (sumc_of_groups != 0).}

\item{c2}{Numeric array.
Construct the row of the time segment factor in which (sumc_of_groups != 0).}

\item{sumc1c2}{Numeric vector
The sum per column of the presence/absence per block rows of the activities-and time segment without zeros.}

\item{sumc1}{Numeric vector.
The sum of the activ_segment.}

\item{levels_sr}{Numeric vector.
levels_sr levels of the sum of rows}

\item{Index_group_number}{Numeric vector.
Index_group_number the group number}

\item{groups}{Numeric vector.
Groups: 4 6 7 3 1 16 2 34 25 82 31 41 21 19 15 27 37 67 77 when Nfactors are two vectors, the time segment and the activities pattern and the RecrID is the only factor for the presence absence of the combination of the activities in a given time segment and groups: 37 1 4 25 when Nfactors is on vector the time segments vector.}

\item{Nfactors}{Numeric vector.
Nfactors: when Nfactors are two vectors, the time segment and the activities pattern then RecrID is the only factor for the presence absence of the combination of the activities in a given time segment and Nfactors is the time segments vector.}

\item{fRecrIDX}{Numeric vector.
fRecrIDX: temp. array for the RecrID vector.}

\item{FRecrID}{Numeric vector.
FRecrID: FRecrID for nRowMatrix is 1 column if the RecrID is the only factor for the presence absence of the combination of the activities in a given time segment (CMATRIX).}

\item{Ngroups}{Numeric vector.
Ngroups is the created vector groupsRecr_at when the RecrID is the only factor for the presence absence of the combination of the activities in a given time segment or the created vector groupsRecr_act_ts when the groupsRecr_at vector and activity pattern are the two factors for the presence absence of the time segment vector.}

\item{fbfactX1}{Numerical array.
fbfactX1:  temp. array for the groupsRecr_at vector.}

\item{fbfactX2}{Numerical array.
fbfactX2:  temp. array for the activity pattern vector}

\item{Fbfact11}{Numerical vector.
Fbfact11: for the groupsRecr_at vector.}

\item{Fbfact22}{Numerical vector.
Fbfact22: for the activity pattern vector}

\item{pos1}{Scalar.
Is an index for slicing an array.}

\item{q}{Scalar.
To find out which combination is.}
}
\value{
list44 Nfactors, Fbfact11, Fbfact22, Ngroups, levels_sr, Index_group_number.
}
\description{
NGroup4 computes the group numbers of recreation areas where all the sequence(s) of their presences (1) on the combination of activities pattern and time segment are four times interrupted by absences (0). Thus in this case we have five groups of one (presences).
}
\details{
NGroup4(nRowMatrix, activ_segment, c1, c2, sumc1c2, sumc1, levels_sr,Index_group_number, groups, Nfactors, fRecrIDX,FRecrID, Ngroups,fbfactX1, fbfactX2, Fbfact11, Fbfact22, pos1, q).
}
