% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wageningen_SQLite_E_2018.R
\docType{data}
\name{WD}
\alias{WD}
\title{Main script for the model ACRER.}
\format{An object of class \code{character} of length 1.}
\usage{
WD
}
\description{
The model ACRER is a recreation model and simulates the number of persons per household and group types, the expected visiting duration as well as the number of visits from a zip code in a city to the destination area.
}
\details{
Several tables are provided in order to compute the abovementioned simulations:
Table_2_RecrID_type_coordinates_prices: 33 records, recreation areas with their Lattitude and Longitude in degrees, their type (eg. sights,park,forest#'/nature area) and the likely entrance prices per recreation area.
Table3_RecrID_activity_pattern: 221 records, Recreation areas ID, recreation areas with their activities pattern.
Table_5_activities_definition_costs: 10 records, activity pattern definition and costs in ??? per activity
Table_14_hgpID_groups_types: 52 records, household group ID, household types definition,group types definition,numer of adults,number of children,presence#' /absence of dog,number of groups and persons.
Table_16_hgpID_min_max_budgets: 52 records, household group ID,household types budget per year,maximum budget (for recreation) per year and maximum budget (for recreation) per day.
Table7_zip_hgpID_budget: 468 records, zip code,household types,group types#',year,numer of adults,number of children,presence/absence of dog,number of groups and persons,maximum budget (for recreation) per year and maximum budget (for recreation) per day.
Table_9_zip_code_RecrID_distances: 891 records,Recreattion ID,zip code, transport type,distances in km, transport travel time in sec and transport speed in km/h.
Table_8_hgpID_distance_costs: 124 records,transport type,household types,group types,maximum of total time (%) per transport type,wait time per full travel, minimum and maximum distance used per transport type,fixed and per km costs per transport type,per person and per km costs per transport type.
Table_10_hgpId_activity_budget: 260 records,household types,group types
,maximum and minimum number of times per activity,maximum and minimum time to spend per activity.
Time_segment_selected: 80 records,Recreation areas ID,number of visits  available per time segment and time available per time segment.
zip_code_coordinates: 9 records, zip codes with their Lattitude and Longitude in degrees.
}
\keyword{datasets}
